+++ region/pom.xml
@@ -70,7 +70,8 @@
                         <Export-Package>
                             org.motechproject.nms.region.service;version=${project.version},
                             org.motechproject.nms.region.repository;version=${project.version},
-                            org.motechproject.nms.region.domain;version=${project.version}
+                            org.motechproject.nms.region.domain;version=${project.version},
+                            org.motechproject.nms.region.csv;version=${project.version},
                         </Export-Package>
                         <Import-Package>
                             net.sf.cglib.core,
