+++ region/src/main/java/org/motechproject/nms/region/controller/LocationDataImportController.java
@@ -4,14 +4,14 @@ import org.motechproject.alerts.contract.AlertService;
 import org.motechproject.alerts.domain.AlertStatus;
 import org.motechproject.alerts.domain.AlertType;
 import org.motechproject.nms.csv.exception.CsvImportException;
-import org.motechproject.nms.region.service.CensusVillageImportService;
-import org.motechproject.nms.region.service.DistrictImportService;
-import org.motechproject.nms.region.service.HealthBlockImportService;
-import org.motechproject.nms.region.service.HealthFacilityImportService;
-import org.motechproject.nms.region.service.HealthSubFacilityImportService;
-import org.motechproject.nms.region.service.LocationDataImportService;
-import org.motechproject.nms.region.service.NonCensusVillageImportService;
-import org.motechproject.nms.region.service.TalukaImportService;
+import org.motechproject.nms.region.csv.CensusVillageImportService;
+import org.motechproject.nms.region.csv.DistrictImportService;
+import org.motechproject.nms.region.csv.HealthBlockImportService;
+import org.motechproject.nms.region.csv.HealthFacilityImportService;
+import org.motechproject.nms.region.csv.HealthSubFacilityImportService;
+import org.motechproject.nms.region.csv.LocationDataImportService;
+import org.motechproject.nms.region.csv.NonCensusVillageImportService;
+import org.motechproject.nms.region.csv.TalukaImportService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
