+++ region/src/main/java/org/motechproject/nms/region/csv/HealthFacilityImportService.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.service;
+package org.motechproject.nms.region.csv;
 
 public interface HealthFacilityImportService extends LocationDataImportService {
 }
