+++ region/src/main/java/org/motechproject/nms/region/csv/LanguageLocationImportService.java
@@ -1,9 +1,9 @@
-package org.motechproject.nms.region.service;
+package org.motechproject.nms.region.csv;
 
 import java.io.IOException;
 import java.io.Reader;
 
-public interface LocationDataImportService {
+public interface LanguageLocationImportService {
 
     void importData(Reader reader) throws IOException;
 }
