+++ region/src/main/java/org/motechproject/nms/region/csv/LocationDataImportService.java
@@ -1,9 +1,9 @@
-package org.motechproject.nms.region.service;
+package org.motechproject.nms.region.csv;
 
 import java.io.IOException;
 import java.io.Reader;
 
-public interface LanguageLocationImportService {
+public interface LocationDataImportService {
 
     void importData(Reader reader) throws IOException;
 }
