+++ region/src/main/java/org/motechproject/nms/region/csv/NonCensusVillageImportService.java
@@ -1,0 +1,6 @@
+package org.motechproject.nms.region.csv;
+
+import org.motechproject.nms.region.csv.LocationDataImportService;
+
+public interface NonCensusVillageImportService extends LocationDataImportService {
+}
