+++ region/src/main/java/org/motechproject/nms/region/csv/TalukaImportService.java
@@ -1,0 +1,6 @@
+package org.motechproject.nms.region.csv;
+
+import org.motechproject.nms.region.csv.LocationDataImportService;
+
+public interface TalukaImportService extends LocationDataImportService {
+}
