+++ region/src/main/java/org/motechproject/nms/region/csv/impl/DistrictImportServiceImpl.java
@@ -1,12 +1,11 @@
-package org.motechproject.nms.region.service.impl.csv;
+package org.motechproject.nms.region.csv.impl;
 
 import org.motechproject.nms.csv.utils.GetLong;
 import org.motechproject.nms.csv.utils.GetString;
 import org.motechproject.nms.region.domain.District;
 import org.motechproject.nms.region.repository.DistrictDataService;
 import org.motechproject.nms.region.repository.StateDataService;
-import org.motechproject.nms.region.service.DistrictImportService;
-import org.motechproject.nms.region.service.impl.csv.BaseLocationImportService;
+import org.motechproject.nms.region.csv.DistrictImportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.ift.CellProcessor;
