+++ region/src/main/java/org/motechproject/nms/region/csv/impl/HealthFacilityImportServiceImpl.java
@@ -1,4 +1,4 @@
-package org.motechproject.nms.region.service.impl.csv;
+package org.motechproject.nms.region.csv.impl;
 
 import org.motechproject.nms.csv.utils.GetInstanceByLong;
 import org.motechproject.nms.csv.utils.GetLong;
@@ -11,9 +11,8 @@ import org.motechproject.nms.region.repository.HealthFacilityTypeDataService;
 import org.motechproject.nms.region.repository.StateDataService;
 import org.motechproject.nms.region.service.DistrictService;
 import org.motechproject.nms.region.service.HealthBlockService;
-import org.motechproject.nms.region.service.HealthFacilityImportService;
+import org.motechproject.nms.region.csv.HealthFacilityImportService;
 import org.motechproject.nms.region.service.TalukaService;
-import org.motechproject.nms.region.service.impl.csv.BaseLocationImportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.ift.CellProcessor;
