+++ region/src/main/java/org/motechproject/nms/region/csv/impl/StateImportServiceImpl.java
@@ -1,11 +1,10 @@
-package org.motechproject.nms.region.service.impl.csv;
+package org.motechproject.nms.region.csv.impl;
 
 import org.motechproject.nms.csv.utils.GetLong;
 import org.motechproject.nms.csv.utils.GetString;
 import org.motechproject.nms.region.domain.State;
 import org.motechproject.nms.region.repository.StateDataService;
-import org.motechproject.nms.region.service.StateImportService;
-import org.motechproject.nms.region.service.impl.csv.BaseLocationImportService;
+import org.motechproject.nms.region.csv.StateImportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.supercsv.cellprocessor.ift.CellProcessor;
