--- region/src/main/java/org/motechproject/nms/region/service/NonCensusVillageImportService.java
@@ -1,4 +1,0 @@
-package org.motechproject.nms.region.service;
-
-public interface NonCensusVillageImportService extends LocationDataImportService {
-}
