+++ region/src/main/resources/META-INF/spring/blueprint.xml
@@ -50,17 +50,17 @@
     <osgi:reference id="circleDataService" interface="org.motechproject.nms.region.repository.CircleDataService" />
 
     <!-- Location Data Import Services -->
-    <osgi:service ref="stateImportService" interface="org.motechproject.nms.region.service.StateImportService" />
-    <osgi:service ref="districtImportService" interface="org.motechproject.nms.region.service.DistrictImportService" />
-    <osgi:service ref="talukaImportService" interface="org.motechproject.nms.region.service.TalukaImportService" />
-    <osgi:service ref="censusVillageImportService" interface="org.motechproject.nms.region.service.CensusVillageImportService" />
-    <osgi:service ref="nonCensusVillageImportService" interface="org.motechproject.nms.region.service.NonCensusVillageImportService" />
-    <osgi:service ref="healthBlockImportService" interface="org.motechproject.nms.region.service.HealthBlockImportService" />
-    <osgi:service ref="healthFacilityImportService" interface="org.motechproject.nms.region.service.HealthFacilityImportService" />
-    <osgi:service ref="healthSubFacilityImportService" interface="org.motechproject.nms.region.service.HealthSubFacilityImportService" />
+    <osgi:service ref="stateImportService" interface="org.motechproject.nms.region.csv.StateImportService" />
+    <osgi:service ref="districtImportService" interface="org.motechproject.nms.region.csv.DistrictImportService" />
+    <osgi:service ref="talukaImportService" interface="org.motechproject.nms.region.csv.TalukaImportService" />
+    <osgi:service ref="censusVillageImportService" interface="org.motechproject.nms.region.csv.CensusVillageImportService" />
+    <osgi:service ref="nonCensusVillageImportService" interface="org.motechproject.nms.region.csv.NonCensusVillageImportService" />
+    <osgi:service ref="healthBlockImportService" interface="org.motechproject.nms.region.csv.HealthBlockImportService" />
+    <osgi:service ref="healthFacilityImportService" interface="org.motechproject.nms.region.csv.HealthFacilityImportService" />
+    <osgi:service ref="healthSubFacilityImportService" interface="org.motechproject.nms.region.csv.HealthSubFacilityImportService" />
 
     <!-- Language Location Codes Import Service -->
-    <osgi:service ref="languageLocationImportService" interface="org.motechproject.nms.region.service.LanguageLocationImportService" />
+    <osgi:service ref="languageLocationImportService" interface="org.motechproject.nms.region.csv.LanguageLocationImportService" />
 
     <!-- Alert Service -->
     <osgi:reference id="alertsService" interface="org.motechproject.alerts.contract.AlertService" />
