+++ testing/src/test/java/org/motechproject/nms/testing/it/region/LocationDataImportServiceBundleIT.java
@@ -18,18 +18,18 @@ import org.motechproject.nms.region.repository.HealthFacilityDataService;
 import org.motechproject.nms.region.repository.HealthFacilityTypeDataService;
 import org.motechproject.nms.region.repository.StateDataService;
 import org.motechproject.nms.region.repository.TalukaDataService;
-import org.motechproject.nms.region.service.CensusVillageImportService;
-import org.motechproject.nms.region.service.DistrictImportService;
+import org.motechproject.nms.region.csv.CensusVillageImportService;
+import org.motechproject.nms.region.csv.DistrictImportService;
 import org.motechproject.nms.region.service.DistrictService;
-import org.motechproject.nms.region.service.HealthBlockImportService;
+import org.motechproject.nms.region.csv.HealthBlockImportService;
 import org.motechproject.nms.region.service.HealthBlockService;
-import org.motechproject.nms.region.service.HealthFacilityImportService;
+import org.motechproject.nms.region.csv.HealthFacilityImportService;
 import org.motechproject.nms.region.service.HealthFacilityService;
-import org.motechproject.nms.region.service.HealthSubFacilityImportService;
+import org.motechproject.nms.region.csv.HealthSubFacilityImportService;
 import org.motechproject.nms.region.service.HealthSubFacilityService;
-import org.motechproject.nms.region.service.NonCensusVillageImportService;
-import org.motechproject.nms.region.service.StateImportService;
-import org.motechproject.nms.region.service.TalukaImportService;
+import org.motechproject.nms.region.csv.NonCensusVillageImportService;
+import org.motechproject.nms.region.csv.StateImportService;
+import org.motechproject.nms.region.csv.TalukaImportService;
 import org.motechproject.nms.region.service.TalukaService;
 import org.motechproject.nms.region.service.VillageService;
 import org.motechproject.nms.testing.service.TestingService;
